backend/requirements.txt
sounddevice>=0.5.2
numpy>=1.24
python-dotenv>=1.0
